requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
apache-airflow==2.6.2
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==6.3.0
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
Babel==2.12.1
black==23.3.0
blinker==1.6.2
cached-property==1.5.2
cachelib==0.10.2
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
fastavro==1.7.4
flake8==5.0.4
Flask==2.2.5
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
numpy==1.21.6
packaging==23.1
pandas==1.3.5
pathspec==0.11.1
peewee==3.16.2
pendulum==2.1.2
platformdirs==3.8.0
prison==0.2.1
psutil==5.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
pyinstrument-cext==0.2.4
PyJWT==2.7.0
PyMySQL==1.1.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-nvd3==0.15.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2019.3
pytzdata==2020.1
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
setproctitle==1.3.2
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.17
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
